+++ neurocheck/requirements-dev.txt
@@ -1,6 +1,9 @@
 -r requirements.txt
 
 # Development Tools (Local Only)
-pytest==7.3.1       # Testing framework
-jupyterlab==4.0.0   # Notebook experiments
-pylint==2.17.4      # Linting
+pytest          # Testing framework
+jupyterlab      # Notebook experiments
+pylint          # Linting
+ipdb            # Interactive debugger
+matplotlib      # Data visualization plots
+seaborn         # Statistical data visualization
