+++ neurocheck/requirements.txt
@@ -2,11 +2,14 @@
 fastapi==0.95.2
 pydantic==1.10.7
 uvicorn==0.22.0
+python-multipart==0.0.6
 
 # ML & Data Handling
 numpy==1.24.3
 pandas==2.0.1
 scikit-learn==1.2.2
-tensorflow==2.12.0
-matplotlib==3.7.1
-seaborn==0.12.2
+xgboost==1.7.6
+mlflow==2.3.1
+
+# CLI / Logs / UX
+colorama==0.4.6
